No_Show_Patience_ElasticBeanstalk/requirements.txt
appnope==0.1.0
attrs==19.1.0
awsebcli==3.15.3
Babel==2.7.0
backcall==0.1.0
bleach==3.1.0
blessed==1.15.0
blist==1.3.6
boto==2.49.0
boto3==1.9.232
botocore==1.12.232
cached-property==1.5.1
cement==2.8.2
certifi==2019.6.16
chardet==3.0.4
Click==7.0
cloudpickle==1.2.2
colorama==0.3.9
colors.py==0.2.2
cycler==0.10.0
darkskylib==0.3.91
dask==2.4.0
decorator==4.4.0
defusedxml==0.6.0
dill==0.3.0
distributed==2.4.0
docker==3.7.3
docker-compose==1.23.2
docker-pycreds==0.4.0
dockerpty==0.4.1
docopt==0.6.2
docutils==0.15.2
entrypoints==0.3
fake-useragent==0.1.11
featuretools==0.10.1
Flask==1.1.1
Flask-Babel==0.12.2
Flask-Table==0.5.0
fsspec==0.5.1
future==0.16.0
graphviz==0.13
HeapDict==1.0.1
hypermax==0.5.0
hyperopt==0.1.2
idna==2.7
ipykernel==5.1.2
ipython==7.8.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
itsdangerous==1.1.0
jedi==0.15.1
Jinja2==2.10.1
jmespath==0.9.4
joblib==0.13.2
json5==0.8.5
jsonschema==2.6.0
jupyter==1.0.0
jupyter-client==5.3.1
jupyter-console==6.0.0
jupyter-core==4.5.0
jupyterlab==1.1.4
jupyterlab-server==1.0.6
kiwisolver==1.1.0
lightgbm==2.3.0
mapq==0.3
MarkupSafe==1.1.1
matplotlib==3.1.1
mistune==0.8.4
msgpack==0.6.1
multiprocess==0.70.8
nbconvert==5.6.0
nbformat==4.4.0
networkx==1.11
notebook==6.0.1
numpy==1.17.2
orderedattrdict==1.5.1
pandas==0.25.1
pandocfilters==1.4.2
panwid==0.2.5
parso==0.5.1
pathspec==0.5.9
patsy==0.5.1
pexpect==4.7.0
pickleshare==0.7.5
praw==6.3.1
prawcore==1.0.1
prometheus-client==0.7.1
prompt-toolkit==2.0.9
psutil==5.6.3
ptyprocess==0.6.0
Pygments==2.4.2
pymongo==3.9.0
pyparsing==2.4.2
pyrsistent==0.15.4
python-dateutil==2.8.0
python-dotenv==0.10.3
pytrends==1.1.3
pytz==2019.2
PyYAML==3.13
pyzmq==18.1.0
qtconsole==4.5.5
raccoon==2.1.5
requests==2.20.1
s3fs==0.3.4
s3transfer==0.2.1
scikit-learn==0.21.3
scipy==1.3.1
seaborn==0.9.0
semantic-version==2.5.0
Send2Trash==1.5.0
simplejson==3.16.0
six==1.11.0
sklearn==0.0
smart-open==1.8.4
sortedcontainers==2.1.0
SQLAlchemy==1.3.8
SQLAlchemy-Utils==0.34.2
statsmodels==0.10.1
tabulate==0.8.3
tblib==1.4.0
termcolor==1.1.0
terminado==0.8.2
testpath==0.4.2
texttable==0.9.1
toolz==0.10.0
tornado==6.0.3
tqdm==4.36.1
traitlets==4.3.2
update-checker==0.16
urllib3==1.24.3
urwid==2.0.1
urwid-utils==0.1.2
virtualenv==16.7.5
wcwidth==0.1.7
webencodings==0.5.1
websocket-client==0.56.0
Werkzeug==0.15.6
widgetsnbextension==3.5.1
xgboost==0.90
yellowbrick==1.0.post1
zict==1.0.0
